SAK-16802 Adjust code in preparation for allowing content size > 2G
git-svn-id: https://source.sakaiproject.org/svn/msgcntr/trunk@65691 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ msgcntr/messageforums-component-impl/src/java/org/sakaiproject/component/app/messageforums/ui/DiscussionForumManagerImpl.java
@@ -2097,7 +2097,7 @@ public class DiscussionForumManagerImpl extends HibernateDaoSupport implements
       attach.setAttachmentName(name);
 
       ContentResource cr = contentHostingService.getResource(attachId);
-      attach.setAttachmentSize((new Integer(cr.getContentLength())).toString());
+      attach.setAttachmentSize((new Long(cr.getContentLength())).toString());
       attach.setCreatedBy(cr.getProperties().getProperty(
           cr.getProperties().getNamePropCreator()));
       attach.setModifiedBy(cr.getProperties().getProperty(
